1. Move Primitive to a separate file. 2. Change all relative import to absolute import. 3. Tweaks some minor problems with context support.
