+++ minpy/array_variants/numpy/numpy_wrapper.py
@@ -1,6 +1,9 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 """Wrapper for NumPy namespace."""
+from __future__ import absolute_import
+from __future__ import print_function
+
 import types
 import numpy as np
