+++ tests/test_context.py
@@ -5,7 +5,7 @@ import minpy.dispatch.policy as policy
 from minpy.context import Context, cpu, gpu, set_context
 
 #np.set_policy(policy.OnlyNumpyPolicy())
-set_context(gpu(1)) # set the global context as gpu(1)
+set_context(gpu(0)) # set the global context as gpu(0)
 
 def sigmoid(x):
     return 0.5 * (np.tanh(x / 2) + 1)
